requirements.txt
amqp==5.0.9
ansicon==1.89.0
arrow==1.2.2
asgiref==3.3.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
blessed==1.19.1
celery==5.2.3
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.1.7
django-celery==3.1.17
django-model-field-meta==0.6
django-oauth-toolkit==1.4.0
django-picklefield==3.0.1
django-q==1.3.9
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jinxed==1.1.0
kombu==5.2.3
MarkupSafe==1.1.1
MouseInfo==0.1.3
mysqlclient==2.1.0
oauthlib==3.1.0
packaging==20.9
paho-mqtt==1.6.1
Pillow==8.3.2
prompt-toolkit==3.0.27
PyAutoGUI==0.9.53
pycryptodome==3.11.0
PyGetWindow==0.0.9
PyJWT==2.1.0
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2021.3
pywhatkit==5.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tuya-connector-python==0.1.2
twilio==7.5.0
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.2.1
wikipedia==1.4.0
wrapt==1.13.3
